Made the tile spawning actually random
At first it was sorta random, but now it is more random. So if you do the same moves twice you won't get the same result...

Still has to be compiled so changes won't apply on current version playable on the 3DS or emulator. If someone compiles it feel free to create a pull request!